modif rotation

+++ lib_fonction_matrix.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   lib_fonction_matrix.c                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: Deydou <Deydou@student.42.fr>              +#+  +:+       +#+        */
+/*   By: alamy <alamy@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/19 13:47:27 by alamy             #+#    #+#             */
-/*   Updated: 2018/01/23 14:42:27 by Deydou           ###   ########.fr       */
+/*   Updated: 2018/01/26 17:08:53 by alamy            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ matrix4_t  matrix_translation_center(t_vecteur4 offset)
 
 /*MATRICE ROTATION*/
 
-matrix4_t  matrix_rotationX(int angle)
+matrix4_t  matrix_rotationX(float angle)
 {
     return matrix4((t_matrix)
 		{1, 0, 0, 0,
@@ -57,7 +57,7 @@ matrix4_t  matrix_rotationX(int angle)
     );
 }
 
-matrix4_t  matrix_rotationY(int angle)
+matrix4_t  matrix_rotationY(float angle)
 {
     return matrix4((t_matrix)
 		{cos(angle), 0, sin(angle), 0,
@@ -67,7 +67,7 @@ matrix4_t  matrix_rotationY(int angle)
     );
 }
 
-matrix4_t  matrix_rotationZ(int angle)
+matrix4_t  matrix_rotationZ(float angle)
 {
     return matrix4((t_matrix)
 		{cos(angle), -sin(angle), 0, 0,
@@ -88,3 +88,13 @@ t_vecteur4 create_vecteur4(int x, int y, int z, int w)
 	return(new_vecteur4);
 }
 
+// t_vecteur4 m4_mult_pos(t_vecteur4 position, matrix4_t matrix) 
+// {
+// 	t_vecteur4 result = create_vecteur4(
+// 		matrix.m[0][0] * position.x1 + matrix.m[0][1] * position.y1 + matrix.m[0][2] * position.z1 + matrix.m[0][3] * position.w1,
+// 		matrix.m[1][0] * position.x1 + matrix.m[1][1] * position.y1 + matrix.m[1][2] * position.z1 + matrix.m[1][3] * position.w1,
+//         matrix.m[2][0] * position.x1 + matrix.m[2][1] * position.y1 + matrix.m[2][2] * position.z1 + matrix.m[2][3] * position.w1,
+//         matrix.m[3][0] * position.x1 + matrix.m[3][1] * position.y1 + matrix.m[3][2] * position.z1 + matrix.m[3][3] * position.w1
+// 	);	
+// 	return result;
+// }
